+++ rog-anime/src/usb.rs
@@ -11,7 +11,7 @@
 use std::str::FromStr;
 
 use dmi_id::DMIID;
-use serde_derive::{Deserialize, Serialize};
+use serde::{Deserialize, Serialize};
 use typeshare::typeshare;
 #[cfg(feature = "dbus")]
 use zbus::zvariant::{OwnedValue, Type, Value};